requirements.txt
Django>=5.1.5,<5.2
djangorestframework>=3.15.2
psycopg2-binary>=2.9.10
python-dotenv>=1.0.1
celery>=5.4.0
redis>=5.0.8
django-celery-beat>=2.7.0
stripe>=10.11.0
requests>=2.32.3
drf-spectacular>=0.27.0
djangorestframework_simplejwt<=5.5